+++ ch_imp/src/evaluation/utils.rs
@@ -8,15 +8,15 @@ use super::{
 const CENTER_DISTANCE_BIT_0: u64 = 0xFF81BDA5A5BD81FF;
 const CENTER_DISTANCE_BIT_1: u64 = 0xFFFFC3C3C3C3FFFF;
 
-pub(super) fn half_board_occupancy_score(pov_occupancy: u64, pov_board: u32, factor: i32) -> i32 {
-    (pov_occupancy & pov_board as u64).count_ones() as i32 * factor
+pub(super) fn half_board_occupancy_score(pov_occupancy: u64, pov_board: u32, factor: i16) -> i16 {
+    (pov_occupancy & pov_board as u64).count_ones() as i16 * factor
 }
 
-pub(super) fn board_occupancy_score(occupancy: u64, board: u64, factor: i32) -> i32 {
-    (occupancy & board as u64).count_ones() as i32 * factor
+pub(super) fn board_occupancy_score(occupancy: u64, board: u64, factor: i16) -> i16 {
+    (occupancy & board as u64).count_ones() as i16 * factor
 }
 
-pub(super) fn piece_positional_reward(occupancy: u64, index: u8, factor: i32) -> i32 {
+pub(super) fn piece_positional_reward(occupancy: u64, index: u8, factor: i16) -> i16 {
     if occupancy.occupied(index) {
         factor
     } else {
@@ -24,7 +24,7 @@ pub(super) fn piece_positional_reward(occupancy: u64, index: u8, factor: i32) ->
     }
 }
 
-pub(super) fn piece_aggregate_score(board: BoardRep, occ: u64, piece_value: PieceValues) -> i32 {
+pub(super) fn piece_aggregate_score(board: BoardRep, occ: u64, piece_value: PieceValues) -> i16 {
     let mut r = 0;
 
     r += board_occupancy_score(board.pawn_bitboard, occ, piece_value[0]);
@@ -36,7 +36,7 @@ pub(super) fn piece_aggregate_score(board: BoardRep, occ: u64, piece_value: Piec
     r
 }
 
-pub(super) fn piece_square_score(piece_bitboard: u64, piece_value_board: PieceValueBoard) -> i32 {
+pub(super) fn piece_square_score(piece_bitboard: u64, piece_value_board: PieceValueBoard) -> i16 {
     let mut bb = piece_bitboard;
     let mut r = 0;
     while bb != 0 {
@@ -50,7 +50,7 @@ pub(super) fn piece_square_score(piece_bitboard: u64, piece_value_board: PieceVa
 pub fn sum_piece_safety_penalties(
     piece_safety_results: &Vec<PieceSafetyInfo>,
     piece_values: PieceValues,
-) -> i32 {
+) -> i16 {
     let mut r = 0;
 
     for &result in piece_safety_results {
@@ -67,14 +67,14 @@ pub fn sum_piece_safety_penalties(
 }
 
 // https://www.chessprogramming.org/Center_Distance
-pub(super) fn distance_to_center(square: u8) -> i32 {
-    (2 * ((CENTER_DISTANCE_BIT_1 >> square) & 1) + ((CENTER_DISTANCE_BIT_0 >> square) & 1)) as i32
+pub(super) fn distance_to_center(square: u8) -> i16 {
+    (2 * ((CENTER_DISTANCE_BIT_1 >> square) & 1) + ((CENTER_DISTANCE_BIT_0 >> square) & 1)) as i16
 }
 
 // https://www.chessprogramming.org/Center_Manhattan-Distance
-pub(super) fn manhattan_distance_to_center(square: u8) -> i32 {
-    let mut file = (square as i32) & 7;
-    let mut rank = (square as i32) >> 3;
+pub(super) fn manhattan_distance_to_center(square: u8) -> i16 {
+    let mut file = (square as i16) & 7;
+    let mut rank = (square as i16) >> 3;
     file ^= (file - 4) >> 8;
     rank ^= (rank - 4) >> 8;
     (file + rank) & 7
